style: move toggle service from chat-member-usecase to chat-room-usecase
