+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/apis/chat/usecase/ChatRoomUseCase.java
@@ -7,7 +7,9 @@ import kr.co.pennyway.api.apis.chat.service.*;
 import kr.co.pennyway.api.common.response.SliceResponseTemplate;
 import kr.co.pennyway.common.annotation.UseCase;
 import kr.co.pennyway.domain.context.chat.dto.ChatRoomDeleteCommand;
+import kr.co.pennyway.domain.context.chat.dto.ChatRoomToggleCommand;
 import kr.co.pennyway.domain.context.chat.service.ChatRoomDeleteService;
+import kr.co.pennyway.domain.context.chat.service.ChatRoomNotificationToggleService;
 import kr.co.pennyway.domain.domains.chatroom.domain.ChatRoom;
 import kr.co.pennyway.domain.domains.chatroom.dto.ChatRoomDetail;
 import lombok.RequiredArgsConstructor;
@@ -25,6 +27,7 @@ public class ChatRoomUseCase {
     private final ChatRoomWithParticipantsSearchService chatRoomWithParticipantsSearchService;
     private final ChatRoomPatchHelper chatRoomPatchHelper;
     private final ChatRoomDeleteService chatRoomDeleteService;
+    private final ChatRoomNotificationToggleService chatRoomNotificationToggleService;
 
     private final ChatMemberSearchService chatMemberSearchService;
 
@@ -74,6 +77,14 @@ public class ChatRoomUseCase {
         ChatRoom chatRoom = chatRoomPatchHelper.updateChatRoom(chatRoomId, request);
 
         return ChatRoomMapper.toChatRoomResDetail(chatRoom, null, true, 1, 0);
+    }
+
+    public void turnOnNotification(Long userId, Long chatRoomId) {
+        chatRoomNotificationToggleService.turnOn(ChatRoomToggleCommand.of(userId, chatRoomId));
+    }
+
+    public void turnOffNotification(Long userId, Long chatRoomId) {
+        chatRoomNotificationToggleService.turnOff(ChatRoomToggleCommand.of(userId, chatRoomId));
     }
 
     public void deleteChatRoom(Long userId, Long chatRoomId) {
